pyproject.toml
[project]
name = "api-certify"
version = "0.1.0"
description = ""
authors = [
    {name = "dinho", email = "claudiosilva.one@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "api-certify"
version = "0.1.0"
description = ""
authors = ["dinho <claudiosilva.one@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = { extras = ["standard"], version = "0.104.1" }
uvicorn = { extras = ["standard"], version = "0.24.0" }
python-dotenv = "1.0.0"
bcrypt = "4.1.2"
motor = "3.3.2"
pymongo = "4.5.0"
pydantic = { extras = ["email"], version = "^2.12.2" }
motor-types = "^1.0.0b4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.0"
pytest = "^8.4.2"
taskipy = "^1.14.1"
pytest-cov = "^4.1.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
format = 'ruff format'
pre_format = 'ruff check --fix'
run = 'uvicorn api_certify.main:app --reload'
pre_test = 'task lint'
test = 'pytest -s -x --cov=api_certify -vv'
post_test = 'coverage html'
kill = 'deactivate && taskkill /F /IM poetry.exe"'
